feat: move context panel to right rail
